Time: N/A (0%), Space: N/A (0%) - LeetHub

+++ 0875-koko-eating-bananas/0875-koko-eating-bananas.cpp
@@ -6,7 +6,7 @@ public:
          int mid;
          int size = piles.size();
        
-         long long count = 0;
+         double count = 0;
          int ans=0;
         
          while(minPile <= maxPile){
